lead-scoring-system/backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.34
psycopg[binary]>=3.2.10
alembic==1.13.2
pydantic>=2.0.0
python-dotenv==1.0.1
redis==5.0.7
tenacity==8.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
slowapi==0.1.9
